requirements.txt
# requirements.txt here provided for information only.
# See `setup.cfg` for dependencies.
black[jupyter]  # Dev only.
build  # Dev only.
cython~=0.29  # Build dependency.
dtaidistance~=2.3
flake8  # Dev only.
importlib-metadata; python_version<"3.8"
isort  # Dev only.
jupyter
matplotlib~=3.5
networkx~=2.8
numpy~=1.22
nvidia-pyindex; python_version == '3.8'  # Benchmarks only.
nvidia-tensorflow[horovod]; python_version == '3.8'  # Benchmarks only.
pandas~=1.4
pre-commit  # Dev only.
pyclustering~=0.10.1
pylint  # Dev only.
pyscaffold  # Dev only.
pytest  # Dev only.
pytest-cov  # Dev only.
scikit_learn~=1.0
scipy~=1.8
setuptools>=43.0.0  # Build dependency.
sympy~=1.10
torch >=1.12, <3.0
tox  # Dev only.
tqdm~=4.64
twine  # Dev only.
